+++ src/Vitorm/Sql/QueryExecutor/Async/ToListAndTotalCountAsync.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 
 using Vit.Linq;
 
+using Vitorm.Sql.DataReader;
 using Vitorm.Sql.SqlTranslate;
 using Vitorm.StreamQuery;
 
@@ -66,7 +67,7 @@ namespace Vitorm.Sql.QueryExecutor
                 // #3 read data
                 {
                     var sql = sqlCount + " ;\r\n" + sqlToList;
-                    using var reader = await dbContext.ExecuteReaderAsync(sql: sql, param: arg.sqlParam, useReadOnly: true);
+                    using var reader = await dbContext.ExecuteReaderAsync(sql: sql, parameters: arg.sqlParam, useReadOnly: true);
                     reader.Read();
                     totalCount = Convert.ToInt32(reader[0]);
                     reader.NextResult();